refactor : 유효성 검사 추가 및 파일이름 변경
