+++ src/pages/FindIdPwPage/FindId.tsx
@@ -3,6 +3,7 @@ import { Form, Field, IData, ErrorMessage } from '../../hooks/useFormState';
 import * as S from '../../components/styles';
 import { useFindIdQuery } from '../../hooks/userQueries';
 import FormControl from '../../components/FormControl';
+import { isValidValue } from '../../utils/Validation';
 
 const FindId = () => {
   const [formIdData, setFormIdData] = useState({
@@ -18,11 +19,11 @@ const FindId = () => {
   const validate = (values: IData<string>) => {
     const errors: IData<string> = {};
 
-    if (!values.userName) {
+    if (!isValidValue(values.userName)) {
       errors.userName = '이름을 입력하세요.';
     }
 
-    if (!values.userEmail) {
+    if (!isValidValue(values.userEmail)) {
       errors.userEmail = '이메일을 입력하세요.';
     }
