+++ src/pages/FindIdPwPage/FindPw.tsx
@@ -3,6 +3,7 @@ import { Form, Field, IData, ErrorMessage } from '../../hooks/useFormState';
 import * as S from '../../components/styles';
 import { useFindPwQuery } from '../../hooks/userQueries';
 import FormControl from '../../components/FormControl';
+import { isValidValue } from '../../utils/Validation';
 
 const FindPw = () => {
   const [formPwData, setFormPwData] = useState({
@@ -23,15 +24,15 @@ const FindPw = () => {
   const validate = (values: IData<string>) => {
     const errors: IData<string> = {};
 
-    if (!values.userName) {
+    if (!isValidValue(values.userName)) {
       errors.userName = '이름을 입력하세요.';
     }
 
-    if (!values.userId) {
+    if (!isValidValue(values.userId)) {
       errors.userId = '아이디를 입력하세요.';
     }
 
-    if (!values.userEmail) {
+    if (!isValidValue(values.userEmail)) {
       errors.userEmail = '이메일을 입력하세요.';
     }
